Добавлен Balance и внесены изменения в Profile

+++ src/main/java/com/finance/app/model/dto/ProfileDto.java
@@ -3,13 +3,11 @@ package com.finance.app.model.dto;
 
 import lombok.Value;
 
-import java.math.BigDecimal;
-
 @Value
 public class ProfileDto {
     Long id;
     String profileName;
-    BigDecimal balance;
+    Long balance;
     Long userId;
 
 }
